+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 + MS-952 Detect user interaction on webview
 + MS-1075 Vungle interstitial video mediation
 + MS-1074 AdColony interstial mediation
++ MS-1076,1080 Vdopia mediation, banner and interstitial
 
 ## RC 2.1
 + MS-857 MS-920 MS-866 AN native response, click handling, cookie management